+++ .github/workflows/pr-highlight-changes.yml
@@ -22,7 +22,7 @@ jobs:
               run: |
                   npm install -g pnpm
                   npm -g i @wordpress/env@5.1.0
-                  pnpm install --filter code-analyzer --filter cli-core
+                  pnpm install --filter code-analyzer --filter monorepo-utils
             - name: Run analyzer
               id: run
               working-directory: tools/code-analyzer
