--- tools/cli-core/README.md
@@ -1,4 +1,0 @@
-### CLI Core
-
-This package contains utilities and libraries providing core functionality to
-cli tools within `tools` so that this functionality can be easily reused.
