--- tools/cli-core/tsconfig.json
@@ -1,7 +1,0 @@
-{
-  "extends": "@tsconfig/node16/tsconfig.json",
-  "ts-node": {
-    "transpileOnly": true,
-    "files": true,
-  }
-}
