+++ tools/code-analyzer/package.json
@@ -9,10 +9,10 @@
 	"dependencies": {
 		"@actions/core": "^1.10.0",
 		"@commander-js/extra-typings": "^0.1.0",
+		"@woocommerce/monorepo-utils": "workspace:*",
 		"@oclif/core": "^2.4.0",
 		"@tsconfig/node16": "^1.0.3",
 		"@types/uuid": "^8.3.4",
-		"cli-core": "workspace:*",
 		"commander": "^9.4.0",
 		"dotenv": "^10.0.0",
 		"simple-git": "^3.10.0",
