+++ tools/monorepo-utils/package.json
@@ -20,7 +20,9 @@
 		"octokit": "^2.0.14",
 		"ora": "^5.4.1",
 		"promptly": "^3.2.0",
-		"semver": "^7.3.2"
+		"semver": "^7.3.2",
+		"simple-git": "^3.10.0",
+		"uuid": "^9.0.0"
 	},
 	"devDependencies": {
 		"@types/jest": "^27.4.1",
