+++ tools/monorepo-utils/src/code-freeze/commands/milestone/utils.ts
@@ -1,24 +1,12 @@
 /**
  * External dependencies
  */
-import { parse, inc } from 'semver';
 import { setOutput } from '@actions/core';
 
 /**
- * Bumps the version according to WP rules.
- *
- * @param {string} version Version to increment
- * @return {string} Incremented version
+ * Internal dependencies
  */
-export const WPIncrement = ( version: string ): string => {
-	const parsedVersion = parse( version );
-	return inc( parsedVersion, parsedVersion.minor === 9 ? 'major' : 'minor' );
-};
-
-const getMajorMinor = ( version: string ) => {
-	const parsedVersion = parse( version );
-	return `${ parsedVersion.major }.${ parsedVersion.minor }`;
-};
+import { getMajorMinor } from '../../../core/version';
 
 /**
  * Set Github outputs.
