+++ tools/monorepo-utils/src/code-freeze/commands/verify-day/index.ts
@@ -2,7 +2,6 @@
  * External dependencies
  */
 import { Command } from '@commander-js/extra-typings';
-import chalk from 'chalk';
 import { setOutput } from '@actions/core';
 
 /**
@@ -14,6 +13,7 @@ import {
 	getToday,
 	getFutureDate,
 } from './utils';
+import { Logger } from '../../../core/logger';
 
 export const verifyDayCommand = new Command( 'verify-day' )
 	.description( 'Verify if today is the code freeze day' )
@@ -28,28 +28,18 @@ export const verifyDayCommand = new Command( 'verify-day' )
 	.action( ( { override, github } ) => {
 		const today = getToday( override );
 		const futureDate = getFutureDate( today );
-		console.log(
-			chalk.yellow( "Today's timestamp UTC is: " + today.toUTCString() )
-		);
-		console.log(
-			chalk.yellow(
-				`Checking to see if ${ DAYS_BETWEEN_CODE_FREEZE_AND_RELEASE } days from today is the second Tuesday of the month.`
-			)
+		Logger.warn( "Today's timestamp UTC is: " + today.toUTCString() );
+		Logger.warn(
+			`Checking to see if ${ DAYS_BETWEEN_CODE_FREEZE_AND_RELEASE } days from today is the second Tuesday of the month.`
 		);
 		const isCodeFreezeDay = isTodayCodeFreezeDay( override );
-		console.log(
-			chalk.green(
-				`${ futureDate.toUTCString() } ${
-					isCodeFreezeDay ? 'is' : 'is not'
-				} release day.`
-			)
+		Logger.notice(
+			`${ futureDate.toUTCString() } ${
+				isCodeFreezeDay ? 'is' : 'is not'
+			} release day.`
 		);
-		console.log(
-			chalk.green(
-				`Today is ${
-					isCodeFreezeDay ? 'indeed' : 'not'
-				} code freeze day.`
-			)
+		Logger.notice(
+			`Today is ${ isCodeFreezeDay ? 'indeed' : 'not' } code freeze day.`
 		);
 
 		if ( github ) {
