+++ tools/monorepo-utils/src/core/environment.ts
