+++ tools/monorepo-utils/src/core/util.ts
@@ -1,0 +1,7 @@
+/**
+ * External dependencies
+ */
+import { promisify } from 'util';
+import { exec } from 'child_process';
+
+export const execAsync = promisify( exec );
