+++ tools/monorepo-utils/src/core/wpenv.ts
@@ -3,60 +3,11 @@
  */
 import { createServer, Server } from 'net';
 import { join } from 'path';
-import { writeFile } from 'fs/promises';
-import { exec } from 'child_process';
-import { promisify } from 'util';
-
-export const execAsync = promisify( exec );
 
 /**
- * Format version string for regex.
- *
- * @param {string} rawVersion Raw version number.
- * @return {string} version regex.
+ * Internal dependencies
  */
-export const getVersionRegex = ( rawVersion: string ): string => {
-	const version = rawVersion.replace( /\./g, '\\.' );
-
-	if ( rawVersion.endsWith( '.0' ) ) {
-		return version + '|' + version.slice( 0, -3 ) + '\\n';
-	}
-
-	return version;
-};
-
-/**
- * Get filename from patch
- *
- * @param {string} str String to extract filename from.
- * @return {string} formatted filename.
- */
-export const getFilename = ( str: string ): string => {
-	return str.replace( /^a(.*)\s.*/, '$1' );
-};
-
-/**
- * Get patches
- *
- * @param {string} content Patch content.
- * @param {RegExp} regex   Regex to find specific patches.
- * @return {string[]} Array of patches.
- */
-export const getPatches = ( content: string, regex: RegExp ): string[] => {
-	const patches = content.split( 'diff --git ' );
-	const changes: string[] = [];
-
-	for ( const p in patches ) {
-		const patch = patches[ p ];
-		const id = patch.match( regex );
-
-		if ( id ) {
-			changes.push( patch );
-		}
-	}
-
-	return changes;
-};
+import { execAsync } from './util';
 
 /**
  * Determine if the default port for wp-env is already taken. If so, see
@@ -148,25 +99,4 @@ export const stopWPEnv = async (
 		}
 		return false;
 	}
-};
-
-/**
- * Generate a JSON file with the data passed.
- *
- * @param  filePath - path to the file to be created.
- * @param  data     - data to be written to the file.
- * @return {Promise<void>} - promise that resolves when the file is written.
- */
-export const generateJSONFile = ( filePath: string, data: unknown ) => {
-	const json = JSON.stringify(
-		data,
-		function replacer( _, value ) {
-			if ( value instanceof Map ) {
-				return Array.from( value.entries() );
-			}
-			return value;
-		},
-		2
-	);
-	return writeFile( filePath, json );
 };
