+++ tools/version-bump/package.json
@@ -16,9 +16,9 @@
 	},
 	"dependencies": {
 		"@commander-js/extra-typings": "^0.1.0",
+		"@woocommerce/monorepo-utils": "workspace:*",
 		"chalk": "^4.1.2",
 		"commander": "9.4.0",
-		"cli-core": "workspace:*",
 		"express": "^4.18.1",
 		"ora": "^5.4.1",
 		"semver": "^7.3.2",
